- extra check for existance of file before delete.

+++ trunk/includes/classes/fileop.class.php
@@ -228,7 +228,7 @@ class fileop {
 			}
 			$d->close();
 			$myreturn=@rmdir($source);
-		} else {
+		} elseif (is_file($source)) {
 			$myreturn=@unlink($source);
 		}
 		return $myreturn;
